backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
email-validator==2.3.0
aiosqlite==0.19.0
greenlet==3.0.3
google-auth==2.27.0
